requirements.txt
asgiref==3.7.2
backoff==2.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
Markdown==3.6
opencage==2.4.0
pillow==10.2.0
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
